+++ x/authz/module/autocli.go
@@ -70,6 +70,13 @@ func (am AppModule) AutoCLIOptions() *autocliv1.ModuleOptions {
 						{ProtoField: "msg_type_url"},
 					},
 				},
+				{
+					RpcMethod: "PruneExpiredGrants",
+					Use:       "prune-grants --from [granter]",
+					Short:     "Prune expired grants",
+					Long:      "Prune up to 75 expired grants in order to reduce the size of the store when the number of expired grants is large.",
+					Example:   fmt.Sprintf(`$ %s tx authz prune-grants --from [mykey]`, version.AppName),
+				},
 			},
 		},
 	}